Finished merging windows-1.3 with release-1.3. A few notes:  * nglib is now dynamic instead of static. It also includes a few patches made for Windows, but also useful for Linux.  * qmake can now be only executed once in the src folder, because it now uses semi-automatic recursion (CONFIG += recursive; SUBDIRS += ...).  * Should still be compatible with the previous build sequence.  * building with mingw(-w64) has not been tested for quite a while now, nor has it with MSVC standard/professional.

+++ src/scripts/build-nglib.sh
@@ -28,9 +28,12 @@ cd netgen_svn
 wget http://engits.eu/files/netgen-4.9.13.zip
 unzip netgen-4.9.13.zip
 rm -f netgen-4.9.13.zip
-rm -rf netgen-mesher
+[ -e netgen-mesher ] && rm -rf netgen-mesher
 mkdir netgen-mesher
 mv netgen-4.9.13 netgen-mesher/netgen
+cd netgen-mesher
+patch -p0 < ../nglib_engrid_mods.diff
+cd ..
 qmake
 make clean
 make -j4
